LATMO-main/main/requirements.txt
chainlit==1.1.404
google_api_python_client==2.135.0
google_auth_oauthlib==1.2.0
langchain==0.2.16
langchain_chroma==0.1.3
langchain_community==0.2.16
langchain_google_genai==1.0.10
langchain_groq==0.1.9
langchain_huggingface==0.0.3
langchain_ollama==0.1.3
langsmith==0.1.116
protobuf==5.28.0
streamlit==1.35.0
pip==24.2
